requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.4
django-crontab==0.7.1
django-db-logger==0.1.12
djangorestframework==3.14.0
drf-api-logger==1.1.14
psycopg2-binary==2.9.7
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
